Refactor executor.rs into focused engine modules
Break up the 2,873-line executor.rs into cohesive modules without
changing any public API or behavior:

- value.rs: Value enum and its methods
- eval.rs: expression evaluator and all operator helpers
- ops_state.rs: state store operations (get, counter, time_since)
- ops_event.rs: event operations (has_active, get_state, in_state)
- ops_semantic.rs: semantic match evaluation
- verdict.rs: RuleVerdict enum and action_to_verdict conversion

executor.rs retains RuleEngine and all integration tests (~1,900 lines).
