+++ crates/rules/rules/src/engine/builtins.rs
@@ -3,7 +3,7 @@
 use chrono::Utc;
 use regex::Regex;
 
-use crate::engine::executor::Value;
+use crate::engine::value::Value;
 use crate::error::RuleError;
 
 /// Dispatch a built-in function call by name.
