+++ crates/rules/rules/src/engine/mod.rs
@@ -1,6 +1,14 @@
 pub mod builtins;
 pub mod context;
+pub mod eval;
 pub mod executor;
+pub mod ops_event;
+pub mod ops_semantic;
+pub mod ops_state;
+pub mod value;
+pub mod verdict;
 
 pub use context::{EmbeddingEvalSupport, EvalContext};
-pub use executor::{RuleEngine, RuleVerdict};
+pub use executor::RuleEngine;
+pub use value::Value;
+pub use verdict::RuleVerdict;
